Improve Android test cleanup safety

+++ src/test/java/tests/android/AndroidCustomWorkoutTest.java
@@ -158,7 +158,8 @@ public class AndroidCustomWorkoutTest extends BaseAndroidTest {
     public void cleanUpCreatedAccount()
             throws InterruptedException {
 
-        if (accountDeletionFlow.canDeleteAccount()) {
+        if (accountDeletionFlow != null
+                && accountDeletionFlow.canDeleteAccount()) {
             accountDeletionFlow.deleteAccount();
         }
     }
